+++ NEWS.md
@@ -2,7 +2,7 @@ DALEXtra 1.3.0
 ----------------------------------------------------------------
 * In `explain_h2o` data parameter will bo converted to data.frame if H2OFrame object was passed.
 * Aspect importance related functions set deprecated. Will be removed with next release.
-
+* `explain_xgboost` function added
 
 DALEXtra 0.2.3
 ----------------------------------------------------------------